LayoutTests/webaudio/getOutputTimestamp-expected.txt

Validate the value returned by AudioContext.getOutputTimestamp()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS contextTime > 0.090 is true
PASS contextTime <= audioContext.currentTime is true
PASS successfullyParsed is true

TEST COMPLETE
